requirements.txt
scipy
snap-stanford
matplotlib
powerlaw
tqdm
networkx
